Extract dependency resolver to a separate class
